Fix restrictive permissions in the workflows file
By default, the workflow runner was only given read only access to the
repository and as a result, a was not able to push to a branch with the
deployed build, which caused all the build errors. Hopefully this should
fix all the problems. What I did was to create a token of my own and set
it as PERSONAL_TOKEN secret in the repository's secrets.